move Chadburn componentent into scenes/game
